Added scaling for Enemies

+++ Battle Arena 3/PlayerStats.cs
@@ -8,9 +8,9 @@ namespace Battle_Arena_3
 {
     class PlayerStats
     {
-        public static int Level { get; set; }
-        public static int XP { get; set; }
-        public static int Gold { get; set; }
+        public static double Level { get; set; }
+        public static double XP { get; set; }
+        public static double Gold { get; set; }
         public static string CharacterName = "default";
